% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCTEOS.R
\name{loadCTEOS}
\alias{loadCTEOS}
\title{Load and Clean CTEOS Data}
\usage{
loadCTEOS(filepath, recode = TRUE)
}
\arguments{
\item{filepath}{Path to your CTEOS file.}

\item{recode}{default as TRUE, if you want to recode variables from the
codebook.}
}
\value{
A CTEOS dataset as a dataset
}
\description{
Load and Clean CTEOS Data
}
